Made it so that load character button displays the character name as well as redirects to battle
Additionally, configured the start game to only have three buttons. Currently the New character button doesn't work 100% correctly. Basically when you make a character with the load game button it will take you to the create character page and when you submit (currently) it goes back to the main menu (TODO: this should redirect to the start game menu)